# hms-openapi-spec.yaml
openapi: 3.0.3
info:
  title: HMS (Hospital Management System) API
  version: "1.0.0"
  description: |
    OpenAPI specification for the HMS endpoints described in the uploaded draft.
    Authentication uses an apiKey-style JWT passed in the Authorization header as "Bearer <token>".
  x-source-file: "/mnt/data/rough-api-2.txt"
servers:
  - url: http://localhost:9000
    description: Local development

# Global security requirement: JWT apiKey in Authorization header
security:
  - JWT: []
paths:
  /hms/id/generate/token:
    post:
      summary: Generate a JWT for a user
      tags: [Users]
      security:
        - JWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCred'
      responses:
        "200":
          description: JWT created successfully
          content:
            application/json:
              schema:
                description: JWT Generation Response
                type: object
                properties:
                    status: 
                      $ref: '#/components/schemas/API-Status'
                    token: 
                      type: string
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/ServerError'
...
  /hms/user/lookup:
    get:
      summary: Lookup user details
      tags: [Users]
      security:
        - JWT: []
      parameters:
        - name: User_ID
          in: query
          schema:
            type: string
        - name: Email_ID
          in: query
          schema:
            type: string
        - name: Phone_Number
          in: query
          schema:
            type: string
        - name: First_Name
          in: query
          schema:
            type: string
        - name: Last_name
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Found users / user details
          content:
            application/json:
              schema:
                description : User Lookup Response
                type: object 
                properties:
                  status: 
                    $ref: '#/components/schemas/API-Status'
                  user_details:
                    type : array
                    items : 
                      $ref: '#/components/schemas/UserResponse'

        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/ServerError'
  /hms/user/update:
    post:
      summary: Update user details
      tags: [Users]
      security:
        - JWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                description: Updated User Details
                type: object
                properties:
                    status: 
                      $ref: '#/components/schemas/API-Status'
                    User_ID:
                      type: string
  
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/ServerError'
  /hms/user/check-password:
    post:
      summary: Check password validity
      tags: [Users]
      security:
        - JWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - User_ID
                - Password
              properties:
                User_ID:
                  type: string
                Password:
                  type: string
      responses:
        "200":
          description: User password check success
          content:
            application/json:
              schema:
                description: Password Check Response
                type: object
                properties:
                    status: 
                      $ref: '#/components/schemas/API-Status'
                    User_ID:
                      type: string
  
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/ServerError'
...
  /hms/appointments/lookup:
    get:
      summary: Lookup appointments
      tags: [Appointments]
      security:
        - JWT: []
      parameters:
        - name: Appointment_ID
          in: query
          schema:
            type: string
        - name: Patient_ID
          in: query
          schema:
            type: string
        - name: Doctor_ID
          in: query
          schema:
            type: string
        - name: Appointment_Date
          in: query
          schema:
            type: string
        - name: Appointment_Status
          in: query
          schema:
            type: string
            enum: [SCHEDULED, COMPLETED, CANCELLED, ALL]
      responses:
        "200":
          description: Appointment(s) returned
          content:
            application/json:
              schema:
                description: Appointment Lookup Response
                type: object
                properties:
                    status: 
                      $ref: '#/components/schemas/API-Status'
                    appointment_details : 
                      type: array 
                      items: 
                        $ref: '#/components/schemas/AppointmentLookup'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/ServerError'
...
  /hms/treatments/lookup:
    get:
      summary: Lookup treatments (history)
      tags: [Treatments]
      security:
        - JWT: []
      parameters:
        - name: Patient_ID
          in: query
          schema:
            type: string
        - name: Doctor_ID
          in: query
          schema:
            type: string
        - name: Appointment_ID
          in: query
          schema:
            type: string
        - name : Appointment_Date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Treatment history returned
          content:
            application/json:
              schema:
                description: Treatment Lookup Response
                type: object 
                properties:
                    status: 
                      $ref: '#/components/schemas/API-Status'
                    treatment_details:
                      type: array
                      items: 
                        $ref: '#/components/schemas/TreatmentLookup'
                  
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/ServerError'
...
  /hms/departments/doctor-lookup: 
    get:
      summary: Lookup doctor-department relations
      tags: [DoctorDept]
      security:
        - JWT: []
      parameters:
        - name: First_Name
          in: query
          schema:
            type: string
        - name: Last_Name
          in: query
          schema:
            type: string
        - name: Speciality
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Relations returned
          content:
            application/json:
              schema:
                description: Doctor-Department Lookup Response
                type: object
                properties:
                    status: 
                      $ref: '#/components/schemas/API-Status'
                    doctordept_details:
                      type: array
                      items: 
                        $ref: '#/components/schemas/DoctorDeptAdd'
               
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/ServerError'
...
  /hms/slots/availability:
    get:
      summary: Find doctors by availability
      tags: [Slots]
      security:
        - JWT: []
      parameters:
        - name: Doctor_ID
          in: query
          schema:
            type: string
        - name: Appointment_Date 
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Doctors matching availability
          content:
            application/json:
              schema:
                description: Doctor Availability Response
                type: object
                properties:
                    status: 
                      $ref: '#/components/schemas/API-Status'   
                    Free_Slots:
                      type: array
                      items: 
                        type: string

        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/ServerError'
components:
  securitySchemes:
    JWT:
      type: apiKey
      in: header
      name: Authorization
      description: JWT token. Include in the Authorization header as  "Authorization <JWT>"
  schemas:
    LoginCred:
      type: object
      required:
        - User_ID 
        - Password
      properties:
        User_ID:
          type: string
        Password:
          type: string
    API-Status:
      type: string
      enum: [success, error]
        

    # Users
...
    UserUpdate:
      type: object
      required: [User_ID]
      properties:
        User_ID:
          type: string
        Email_ID:
          type: string
        Sex:
          type: string
          enum: [M, F, O]
        First_Name:
          type: string
        Last_Name:
          type: string
        Phone_Number:
          type: string
        User_Profile:
          type: object
        Password: 
          type: string
    UserResponse:
      type: object
      properties:
        User_ID:
          type: string
        Email_ID:
          type: string
        Sex:
          type: string
        First_Name:
          type: string
        Last_Name:
          type: string
        User_Type:
          type: string
        Phone_Number:
          type: string
        User_Profile:
          type: object

    # Appointments
...
    AppointmentLookup:
      type: object
      required:
        - Patient_ID
        - Doctor_ID
        - Appointment_Date
        - Appointment_Time
      properties:
        Patient_ID:
          type: string
        patient_First_Name:
          type: string
        patient_Last_Name:
          type: string
        Doctor_ID:
          type: string
        doctor_First_Name:
          type: string
        doctor_Last_Name:
          type: string
        Specialities:
          type: string
        Appointment_Date:
          type: string
          description: "Date in DD/MM/YY or ISO format"
        Appointment_Time:
          type: string
          description: "Time in HH:MM"
        Appointment_Status:
          type: string
          enum: [SCHEDULED, COMPLETED, CANCELLED]

    
    # Treatments
...
    TreatmentLookup:
      type: object
      required:
        - Appointment_ID
        - Notes
      properties:
        Appointment_ID:
          type: string
        Appointment_Date:
          type: string
        Diagnosis:
          type: string
        Prescription:
          type: string
        Notes:
          type: string
        Doctor_ID:
          type: string
        doctor_First_Name:
          type: string
        doctor_Last_Name:
          type: string
        Patient_ID:
          type: string
        patient_First_Name:
          type: string
        patient_Last_Name:
          type: string
        Specialities:
          type: string

    # Departments & Doctor_Dept
...
    DoctorDeptAdd:
      type: object
      required: [Doctor_ID, Dept_ID,Dept_Position]
      properties:
        Doctor_ID:
          type: string
        Dept_ID:
          type: string
        Dept_Position: 
          type: string 
          enum: ['HOD','CONSULTANT','RESIDENT']
  
    # Slots
...
    Error:
      type: object
      properties:
        status : 
            $ref: '#/components/schemas/API-Status'
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request - validation / missing parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - role not permitted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...